+++ engine/src/test/java/io/camunda/zeebe/engine/processing/bpmn/subprocess/InterruptingEventSubprocessConcurrencyTest.java
@@ -16,7 +16,6 @@ import io.camunda.zeebe.model.bpmn.Bpmn;
 import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
 import io.camunda.zeebe.model.bpmn.builder.ProcessBuilder;
 import io.camunda.zeebe.protocol.impl.record.value.message.MessageRecord;
-import io.camunda.zeebe.protocol.record.intent.IncidentIntent;
 import io.camunda.zeebe.protocol.record.intent.MessageIntent;
 import io.camunda.zeebe.protocol.record.intent.MessageSubscriptionIntent;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
@@ -89,127 +88,6 @@ public class InterruptingEventSubprocessConcurrencyTest {
             .processMessageSubscription(
                 ProcessMessageSubscriptionIntent.CORRELATE,
                 eventSubprocessSubscriptionCreated.getValue()));
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords()
-                .withProcessInstanceKey(processInstanceKey)
-                .limitToProcessInstanceCompleted())
-        .extracting(r -> tuple(r.getValue().getBpmnElementType(), r.getIntent()))
-        .containsSubsequence(
-            tuple(BpmnElementType.EVENT_SUB_PROCESS, ProcessInstanceIntent.ELEMENT_ACTIVATED),
-            tuple(BpmnElementType.EVENT_SUB_PROCESS, ProcessInstanceIntent.ELEMENT_COMPLETED),
-            tuple(BpmnElementType.PROCESS, ProcessInstanceIntent.ELEMENT_COMPLETED));
-  }
-
-  @Test
-  // https://github.com/camunda/zeebe/issues/6565
-  public void shouldEndProcessWithParallelFlow() {
-    // given
-    final ProcessBuilder process = Bpmn.createExecutableProcess(PROCESS_ID);
-
-    process
-        .eventSubProcess("event_sub_proc")
-        .startEvent("event_sub_start")
-        .interrupting(true)
-        .message(b -> b.name(MSG_NAME).zeebeCorrelationKeyExpression("key"))
-        .endEvent("event_sub_end");
-
-    final BpmnModelInstance model =
-        process
-            .startEvent("start_proc")
-            .sequenceFlowId("toParallel")
-            .parallelGateway("parallel")
-            .intermediateCatchEvent("catch")
-            .message(m -> m.name("msg").zeebeCorrelationKeyExpression("key"))
-            .endEvent("end_proc")
-            .moveToLastGateway()
-            .intermediateCatchEvent("catch1")
-            .message(m -> m.name("msg1").zeebeCorrelationKeyExpression("key1"))
-            .endEvent("end_proc1")
-            .done();
-
-    engineRule.deployment().withXmlResource(model).deploy();
-
-    final long processInstanceKey =
-        engineRule
-            .processInstance()
-            .ofBpmnProcessId(PROCESS_ID)
-            .withVariables(Map.of("key", 123, "key1", 123))
-            .create();
-
-    // when
-    RecordingExporter.messageSubscriptionRecords(MessageSubscriptionIntent.CREATED)
-        .withProcessInstanceKey(processInstanceKey)
-        .withMessageName(MSG_NAME)
-        .await();
-    RecordingExporter.processInstanceRecords()
-        .withElementType(BpmnElementType.START_EVENT)
-        .withIntent(ProcessInstanceIntent.ELEMENT_ACTIVATING)
-        .withElementId("start_proc")
-        .withProcessInstanceKey(processInstanceKey)
-        .await();
-
-    engineRule
-        .message()
-        .withName(MSG_NAME)
-        .withCorrelationKey("123")
-        .withVariables(Map.of("key", "123"))
-        .publish();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords()
-                .withProcessInstanceKey(processInstanceKey)
-                .limitToProcessInstanceCompleted())
-        .extracting(r -> tuple(r.getValue().getBpmnElementType(), r.getIntent()))
-        .containsSubsequence(
-            tuple(BpmnElementType.EVENT_SUB_PROCESS, ProcessInstanceIntent.ELEMENT_ACTIVATED),
-            tuple(BpmnElementType.EVENT_SUB_PROCESS, ProcessInstanceIntent.ELEMENT_COMPLETED),
-            tuple(BpmnElementType.PROCESS, ProcessInstanceIntent.ELEMENT_COMPLETED));
-  }
-
-  @Test
-  public void shouldTerminateXorWithIncident() {
-    // given
-    final ProcessBuilder process = Bpmn.createExecutableProcess(PROCESS_ID);
-
-    process
-        .eventSubProcess("event_sub_proc")
-        .startEvent("event_sub_start")
-        .interrupting(true)
-        .message(b -> b.name(MSG_NAME).zeebeCorrelationKeyExpression("key"))
-        .endEvent("event_sub_end");
-
-    final BpmnModelInstance model =
-        process
-            .startEvent("start_proc")
-            .sequenceFlowId("toXor")
-            .exclusiveGateway("xor")
-            .condition("=yolo")
-            .endEvent()
-            .done();
-
-    engineRule.deployment().withXmlResource(model).deploy();
-
-    final long processInstanceKey =
-        engineRule
-            .processInstance()
-            .ofBpmnProcessId(PROCESS_ID)
-            .withVariables(Map.of("key", 123, "key1", 123))
-            .create();
-
-    // when
-    RecordingExporter.incidentRecords(IncidentIntent.CREATED)
-        .withProcessInstanceKey(processInstanceKey)
-        .await();
-
-    engineRule
-        .message()
-        .withName(MSG_NAME)
-        .withCorrelationKey("123")
-        .withVariables(Map.of("key", "123"))
-        .publish();
 
     // then
     assertThat(
